PageRank Optimem : Edited normalization
Before, the normalization was done to have all the vector elements in the range [0 ; 1] by dividing each element by the max value of the vector. The sum of the vector elements after normalization was not 1.
Now, the normalization is done in order to have the sum equal to 1 after normalization. Each element is divided by the "sum before normalization"